requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
scikit-learn>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pillow>=10.0.0
python-multipart>=0.0.9
